finish 4.a

+++ include/img_process.hpp
@@ -27,9 +27,9 @@ Mat padding(Mat oldimage,Mat padimage,int padsize,int resize,int mode);
 
 unsigned char* storeMat2Pixel(Mat image,unsigned char* pixel,int size);
 
-Mat conv(Mat paded_image, Mat out_image, int* mask, int image_length, int mask_length);
+Mat conv(Mat paded_image, Mat out_image, double* mask, int image_length, int mask_length);
 
-int* highboostMask(int* mask,int A,int masksize);
+double* highboostMask(double* mask,int A,int masksize);
 
 Mat localEnhencement(Mat padedimage, int imglength, int masklength);
